% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_fast.R
\name{colnames_fast}
\alias{colnames_fast}
\title{Retrieve column names of a file}
\usage{
colnames_fast(file, separator = "\\t")
}
\arguments{
\item{file}{Filename to retrieve colnames from.}

\item{separator}{The column delimiter of the file.}
}
\value{
Returns the column names as a character vector.
}
\description{
Returns the column names of a file, provided these are the first row of the file.
}
