% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_columns.R
\name{combine_columns}
\alias{combine_columns}
\title{Create a new column by combining two existing columns}
\usage{
combine_columns(
  file,
  column_1 = 1,
  column_2 = 2,
  combination_symbol = ":",
  separator = "\\t",
  add_to_file = TRUE
)
}
\arguments{
\item{file}{Filename.}

\item{column_1}{Column 1 to be combined (left side).}

\item{column_2}{Column 2 to be combined (right side).}

\item{separator}{The column delimiter of the file.}

\item{add_to_file}{If TRUE, appends the new column to the file before returning the file.}
}
\value{
Returns either the new column or the file as matrix, with new column appended.
}
\description{
Returns a combined set of columns as one column, or the original file with the combined column appended.
}
