% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_fast.R
\name{dim_fast}
\alias{dim_fast}
\title{Return the dimensions of a file without reading it in}
\usage{
dim_fast(file, separator = "\\t")
}
\arguments{
\item{file}{Filename to retrieve dimensions of.}

\item{separator}{The column delimiter of the file.}
}
\value{
Returns the dimensions of a file as numeric vector (row, column).
}
\description{
Returns the row, column dimensions of a file (using awk in linux).
}
