% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_fast.R
\name{head_fast}
\alias{head_fast}
\title{Show first lines of file without reading it in}
\usage{
head_fast(file, lines = 10, separator = "\\t")
}
\arguments{
\item{file}{Filename to print first lines of.}

\item{lines}{Number of rows to print.}

\item{separator}{The column delimiter of the file.}
}
\value{
Returns the first few rows of the file as a matrix.
}
\description{
Passes 'head' function to the command line (linux).
}
