% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fast.R
\name{read_fast}
\alias{read_fast}
\title{Read in a file using linux}
\usage{
read_fast(file, separator = "\\t")
}
\arguments{
\item{file}{Filename of file to load.}

\item{separator}{The column delimiter of the file.}
}
\value{
Returns a matrix.
}
\description{
Loads a file as matrix using cat in linux.
}
