% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_rows.R
\name{sample_rows}
\alias{sample_rows}
\title{Randomly sample rows from a file without first loading it}
\usage{
sample_rows(
  file,
  no_of_rows = 10,
  exclude_condition = NULL,
  exclude_condition_col = 1,
  separator = "\\t"
)
}
\arguments{
\item{file}{Filename to retrieve rows from.}

\item{no_of_rows}{Number of rows to sample.}

\item{exclude_condition}{Rows not matching the condition will not be sampled from.}

\item{exclude_condition_col}{Column to which the exclusion condition should be applied.}

\item{separator}{The column delimiter of the file.}
}
\value{
Returns a random subset of specified size from a file as a matrix.
}
\description{
Returns a random subset of arbitrary size from a file (using awk and shuf in linux).
}
