% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_fast.R
\name{sort_fast}
\alias{sort_fast}
\title{Quickly sort a file and return}
\usage{
sort_fast(file, column = 1, descending = TRUE, separator = "\\t")
}
\arguments{
\item{file}{Filename to sort.}

\item{column}{Column to sort by.}

\item{descending}{If TRUE, sorts in descending order.}

\item{separator}{The column delimiter of the file.}
}
\value{
Returns the sorted file as matrix.
}
\description{
Returns the file sorted by a column.
}
