% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_condition.R
\name{subset_condition}
\alias{subset_condition}
\title{Subset a file by a logical condition on a column}
\usage{
subset_condition(file, column = 10, condition = "< 5e-8", separator = "\\t")
}
\arguments{
\item{file}{Filename to retrieve subset of.}

\item{column}{Column to apply logical condition to.}

\item{condition}{Logical condition to apply.}

\item{separator}{The column delimiter of the file.}
}
\value{
Returns a subset of a file based on the condition.
}
\description{
Returns the subset of a file matching a given logical condition as a matrix (with awk in linux).
}
