% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_grep.R
\name{subset_grep}
\alias{subset_grep}
\title{Subset a file based on grepping}
\usage{
subset_grep(
  file,
  column = 1,
  pattern = NULL,
  separator = "\\t",
  grep = TRUE,
  first_row = TRUE
)
}
\arguments{
\item{file}{Filename to grep in.}

\item{column}{Column to apply the grep to.}

\item{pattern}{A regular expression.}

\item{separator}{The column delimiter of the file.}

\item{grep}{If FALSE, does not grep but exactly matches the pattern.}

\item{first_row}{If TRUE, returns first row as well.}
}
\value{
Returns a subset of a file as a matrix.
}
\description{
Retrieves a subset of a file based on grepping for a pattern (using grep/awk in linux).
}
