% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail_fast.R
\name{tail_fast}
\alias{tail_fast}
\title{Show last lines of file without reading it in}
\usage{
tail_fast(file, lines = 10, separator = "\\t")
}
\arguments{
\item{file}{Filename to print last lines of.}

\item{lines}{Number of rows to print.}

\item{separator}{The column delimiter of the file.}
}
\value{
Returns the last few rows of the file as a matrix.
}
\description{
Passes 'tail' function to the command line (linux).
}
