% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_match.R
\name{subset_match}
\alias{subset_match}
\title{Subset a file by exact matching a set of identifiers}
\usage{
subset_match(file, column = 1, ids = NULL, separator = "\\t", first_row = TRUE)
}
\arguments{
\item{file}{Filename to subset.}

\item{column}{Column to identify exact matches in.}

\item{ids}{Identifiers (a character vector).}

\item{separator}{The column delimiter of the file.}

\item{first_row}{If TRUE, returns first row as well.}
}
\value{
Returns a subset of the file as a matrix.
}
\description{
Returns a subset of a file based on exactly matching a set of identifiers (using awk in linux).
}
