% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_match_fast.R
\name{subset_match_fast}
\alias{subset_match_fast}
\title{Subsetting based on an exact match with identifiers by hash-based lookup}
\usage{
subset_match_fast(file, column = 1, ids = NULL, separator = "\\t")
}
\arguments{
\item{file}{Filename to retrieve dimensions of.}

\item{column}{Column to check the exact matches in.}

\item{ids}{Identifiers to match (character vector).}

\item{separator}{The column delimiter of the file.}
}
\value{
Returns a subset of the file as a matrix.
}
\description{
Subsets a file by writing identifiers to a separate file, then matching the original file against this (using awk in linux).
}
